Add location functionality to map
This adds all of the code necessary to get location updates, get the last known location, and get the location from a saved state within our Maps Activity. Currently this can only be tested using the emulator (RPI3 having location problems), and in order to change the location on the default emulator from android studio you have to go into the device settings (three little dots at the bottom of the sidebar of the emulator) and then choose Location and change the lat/lng values and the map should update (on an interval of 5 seconds for refresh).