Update UI
Moves the app bar to a cardview floating above the map, as well as adds space to the right of the map for widgets or whatever else we intend to add. Also chanegs values to px instead of dp since we are building for a single display (which will fix scaling disparities between emulator and RPI3).